chore(client): make /memories v2 canonical; delete v1 mesh-driven page
The legacy mesh-driven memories.page.tsx coexisted with the
curation-focused memories-v2.page.tsx at /memories and /memories/v2.
v2 is the production list view; v1 was demoware. Renamed v2 → memories.page.tsx,
renamed export MemoriesV2 → Memories, redirected /memories/v2 → /memories.

The 3D mesh visualization is preserved in components/memories/mesh
and reachable via /mesh-showcase; surfacing it as a tab inside
/memories is left to follow-up product work.
